[Minor] Fused experts refactor (#15914)
